Split charts.py into multiple files by function type for adults, refs #7
