+++ processdata.py
@@ -1,0 +1,182 @@
+import pandas as pd
+import numpy as np
+import math
+import matplotlib.pyplot as plt
+import matplotlib as mpl
+from IPython.display import FileLink, FileLinks, Markdown
+
+def setup_individual_obs_df(obs_df):
+  obs_df.rename(columns={'result':'clean_value'}, inplace=True)
+  obs_df['clean_cat'] = obs_df['clean_value'].astype('category')
+  obs_df['age'] = obs_df['age_years'] #obs_df['agedays'] / 365
+  obs_df['include'] = obs_df.clean_value.eq("Include")
+  return obs_df
+
+def setup_percentiles(percentiles):
+  # expand decade rows into one row per year
+  pct = percentiles[percentiles['Age (All race and Hispanic-origin groups)'] != '20 and over'].copy()
+  pct.loc[pct['Age_low'] == 20, 'Age_low'] = 18
+  range_col = pct.apply(lambda row: row.Age_high - row.Age_low + 1, axis=1)
+  pct = pct.assign(range=range_col.values)
+  dta = pd.DataFrame((np.repeat(pct.values, pct['range'], axis=0)), columns=pct.columns)
+  dta['count'] = dta.groupby(['Sex', 'Measure', 'Age_low', 'Age_high']).cumcount() # ['range'].transform('cumcount')
+  dta['age_years'] = dta['Age_low'] + dta['count']
+  # add standard deviation and other values
+  dta['sqrt'] = np.sqrt(pd.to_numeric(dta['Number of examined persons']))
+  dta['sd'] = dta['Standard error of the mean'] * dta['sqrt']
+  dta['sex'] = np.where(dta['Sex'] == 'Male', 0, 1)
+  dta.rename(columns={'Measure':'param'}, inplace=True)
+  dta.drop(columns=['Age (All race and Hispanic-origin groups)', 'Sex', 'Age_low', 'sqrt', 'Standard error of the mean',
+                  'Age_high', 'range', 'count', 'Number of examined persons'], inplace=True)
+  # smooth percentiles between X9-(X+1)1 (i.e., 29-31)
+  dta['decade'] = np.where(dta['age_years'] == (round(dta['age_years'].astype(float), -1)), 1, 0)
+  mcol_list = ['Mean', 'sd', 'P5', 'P10', 'P15', 'P25', 'P50', 'P75', 'P85', 'P90', 'P95']
+  for col in mcol_list:
+    dta[col] = np.where((dta['decade'] == 1) & (dta['age_years'] < 110), (dta[col] + dta[col].shift(1))/2, dta[col])
+  dta.drop(columns={'decade'}, inplace=True)
+  col_list = ['param', 'sex', 'age_years'] + mcol_list
+  dta = dta.reindex(columns=col_list)
+  return dta
+
+def keep_age_range(df):
+  obs_grp = df
+  def label_excl_grp(row):
+    if row['age'] < 20: return ' Below 20 (Exclude)'
+    if (row['age'] >= 20) & (row['age'] < 30): return ' Between 20 and 30'
+    if (row['age'] >= 30) & (row['age'] < 40): return ' Between 30 and 40'
+    if (row['age'] >= 40) & (row['age'] < 50): return ' Between 40 and 50'
+    if (row['age'] >= 50) & (row['age'] < 60): return ' Between 50 and 60'
+    if (row['age'] >= 60) & (row['age'] < 65): return ' Between 60 and 65'
+    if row['age'] > 65: return 'Above 65 (Exclude)'
+  label_excl_col = obs_grp.apply(lambda row: label_excl_grp(row), axis=1)
+  obs_grp = obs_grp.assign(cat=label_excl_col.values)
+  obs_grp = obs_grp.groupby('cat')['subjid'].count().reset_index().sort_values('cat', ascending=True)
+  colors = ['C3', 'C0', 'C0', 'C0', 'C0', 'C0', 'C3']
+  obs_grp_plot = obs_grp.plot.bar('cat', 'subjid', color=colors, legend=None)
+  obs_grp_plot.set_xlabel('')
+  plt.xticks(rotation=45, ha='right')
+  obs_grp_plot.yaxis.set_major_formatter(mpl.ticker.StrMethodFormatter('{x:,.0f}'))
+  return df[df['age'].between(20, 65, inclusive=True)]
+
+def setup_merged_df(obs_df):
+  obs_df = obs_df.assign(height=obs_df['measurement'], weight=obs_df['measurement'])
+  obs_df.loc[obs_df.param == 'WEIGHTKG', 'height'] = np.NaN 
+  obs_df.loc[obs_df.param == 'HEIGHTCM', 'weight'] = np.NaN
+  heights = obs_df[obs_df.param == 'HEIGHTCM']
+  weights = obs_df[obs_df.param == 'WEIGHTKG']
+  merged = heights.merge(weights, on=['subjid', 'age_years', 'sex'], how='outer')
+  only_needed_columns = merged.drop(columns=['param_x', 'measurement_x', 'clean_value_x',
+                                             'age_x', 'weight_x', 'id_y', 'param_y',
+                                             'measurement_y', 'clean_value_y', 'height_y', 'reason_x'])
+  clean_column_names = only_needed_columns.rename(columns={'clean_cat_x': 'height_cat',
+                                                           'include_x': 'include_height',
+                                                           'height_x': 'height',
+                                                           'clean_cat_y': 'weight_cat',
+                                                           'age_y': 'age',
+                                                           'include_y': 'include_weight',
+                                                           'weight_y': 'weight',
+                                                           'age_years_y':'age_years', 
+                                                           'reason_y':'reason', 
+                                                           'id_x':'id'})
+  clean_column_names['bmi'] = clean_column_names['weight'] / ((clean_column_names['height'] / 100) ** 2)
+  clean_column_names['rounded_age'] = np.around(clean_column_names.age)
+  clean_column_names['include_both'] = clean_column_names['include_height'] & clean_column_names['include_weight']
+  return clean_column_names
+
+def setup_bmi(merged_df, obs):
+  data = merged_df[['id', 'subjid', 'sex', 'age_years', 'age', 'rounded_age', 'bmi', 'weight_cat', 'height_cat', 'include_both']]
+  def label_incl(row):
+    if (row['include_both'] == True): return 'Include'
+    elif (row['weight_cat'] == 'Implausible') | (row['height_cat'] == 'Implausible'): return 'Implausible'
+    else: return 'Only Wt or Ht'
+  incl_col = data.apply(lambda row: label_incl(row), axis=1)
+  data = data.assign(clean_cat=incl_col.values)
+  data['param'] = 'BMI'
+  data.rename(columns={'bmi':'measurement'}, inplace=True)
+  return pd.concat([obs, data])
+
+def data_frame_names(da_locals):
+  frames = []
+  for key, value in da_locals.items():
+    if isinstance(value, pd.DataFrame):
+      frames.append(key)
+  return frames
+
+def exclusion_information(obs):
+  """
+  Provides a count and percentage of growthcleanr categories by measurement type (param).
+
+  Parameters:
+  obs: a DataFrame, in the format output by setup_individual_obs_df
+
+  Returns:
+  A DataFrame with the counts and percentages
+  """
+  exc = obs.groupby(['param', 'clean_cat']).agg({'id': 'count'}).reset_index().pivot(index="clean_cat", columns='param', values='id')
+  exc['height percent'] = exc['HEIGHTCM'] / exc['HEIGHTCM'].sum() * 100
+  exc['weight percent'] = exc['WEIGHTKG'] / exc['WEIGHTKG'].sum() * 100
+  exc = exc.fillna(0)
+  exc['total'] = exc['HEIGHTCM'] + exc['WEIGHTKG']
+  exc = exc[['HEIGHTCM', 'height percent', 'WEIGHTKG', 'weight percent', 'total']]
+  exc = exc.sort_values('total', ascending=False)
+  return exc.style.format({'HEIGHTCM': "{:.0f}".format, 'height percent': "{:.2f}%",
+                           'WEIGHTKG': "{:.0f}".format, 'weight percent': "{:.2f}%"})
+
+def export_to_csv(da_locals, selection_widget, out):
+  df_name = selection_widget.value
+  da_locals[df_name].to_csv('growthviz-data/output/{}.csv'.format(df_name), index=False)
+  out.clear_output()
+  out.append_display_data(FileLinks('growthviz-data/output'))
+
+def clean_swapped_values(merged_df):
+  """
+  This function will look in a DataFrame for rows where the height_cat and weight_cat are set to
+  "Swapped-Measurements". It will then swap the height and weight values for those rows.
+  It will also create two new columns: postprocess_height_cat and postprocess_weight_cat.
+  The values for these columns is copied from the original categories except in the case where
+  swaps are fixed when it is set to "Include-Fixed-Swap".
+
+  Parameters:
+  merged_df: (DataFrame) with subjid, height, weight, include_height and include_weight columns
+
+  Returns:
+  The cleaned DataFrame
+  """
+  merged_df['postprocess_height_cat'] = merged_df['height_cat']
+  merged_df['postprocess_height_cat'] = merged_df['postprocess_height_cat'].cat.add_categories(['Include-Fixed-Swap'])
+  merged_df['postprocess_weight_cat'] = merged_df['weight_cat']
+  merged_df['postprocess_weight_cat'] = merged_df['postprocess_weight_cat'].cat.add_categories(['Include-Fixed-Swap'])
+  merged_df.loc[merged_df['height_cat'] == 'Swapped-Measurements', ['height', 'weight']] = merged_df.loc[merged_df['height_cat'] == 'Swapped-Measurements', ['weight', 'height']].values
+  merged_df.loc[merged_df['height_cat'] == 'Swapped-Measurements', 'postprocess_height_cat'] = 'Include-Fixed-Swap'
+  merged_df.loc[merged_df['weight_cat'] == 'Swapped-Measurements', 'postprocess_weight_cat'] = 'Include-Fixed-Swap'
+  merged_df['bmi'] = merged_df['weight'] / ((merged_df['height'] / 100) ** 2)
+  return merged_df
+
+def clean_unit_errors(merged_df):
+  """
+  This function will look in a DataFrame for rows where the height_cat and weight_cat are set to
+  "Unit-Error-High" or "Unit-Error-Low". It will then multiply / divide the height and weight values to convert them.
+  It will also create two new columns: postprocess_height_cat and postprocess_weight_cat.
+  The values for these columns are copied from the original categories except in the case where
+  unit errors are fixed when it is set to "Include-UH" or "Include-UL" respectively.
+
+  Parameters:
+  merged_df: (DataFrame) with subjid, height, weight, include_height and include_weight columns
+
+  Returns:
+  The cleaned DataFrame
+  """
+  merged_df['postprocess_height_cat'] = merged_df['height_cat']
+  merged_df['postprocess_height_cat'] = merged_df['postprocess_height_cat'].cat.add_categories(['Include-UH','Include-UL'])
+  merged_df['postprocess_weight_cat'] = merged_df['weight_cat']
+  merged_df['postprocess_weight_cat'] = merged_df['postprocess_weight_cat'].cat.add_categories(['Include-UH','Include-UL'])
+  merged_df.loc[merged_df['height_cat'] == 'Unit-Error-Low', 'height'] = (merged_df.loc[merged_df['height_cat'] == 'Unit-Error-Low', 'height'] * 2.54)
+  merged_df.loc[merged_df['height_cat'] == 'Unit-Error-High', 'height'] = (merged_df.loc[merged_df['height_cat'] == 'Unit-Error-High', 'height'] / 2.54)
+  merged_df.loc[merged_df['weight_cat'] == 'Unit-Error-Low', 'weight'] = (merged_df.loc[merged_df['weight_cat'] == 'Unit-Error-Low', 'weight'] * 2.2046)
+  merged_df.loc[merged_df['weight_cat'] == 'Unit-Error-High', 'weight'] = (merged_df.loc[merged_df['weight_cat'] == 'Unit-Error-High', 'weight'] / 2.2046)
+  merged_df.loc[merged_df['height_cat'] == 'Unit-Error-Low', 'postprocess_height_cat'] = 'Include-UL'
+  merged_df.loc[merged_df['height_cat'] == 'Unit-Error-High', 'postprocess_height_cat'] = 'Include-UH'
+  merged_df.loc[merged_df['weight_cat'] == 'Unit-Error-Low', 'postprocess_weight_cat'] = 'Include-UL'
+  merged_df.loc[merged_df['weight_cat'] == 'Unit-Error-High', 'postprocess_weight_cat'] = 'Include-UH'
+  merged_df['bmi'] = merged_df['weight'] / ((merged_df['height'] / 100) ** 2)
+  return merged_df
